Rename files in Baja (session manager) module
