+++ session-manager/src/lib.rs
@@ -19,9 +19,9 @@
 #[macro_use]
 extern crate sgx_tstd as std;
 
-pub mod baja;
-pub use self::baja::*;
-pub mod baja_session;
-pub use self::baja_session::*;
+pub mod session_context;
+pub use self::session_context::*;
+pub mod session;
+pub use self::session::*;
 pub mod error;
 pub use self::error::BajaError;
