+++ session-manager/src/session.rs
